backend/requirements.txt
django==4.1.4
djangorestframework==3.14.0
environs==9.5.0
Pillow==9.3.0
django-cleanup==6.0.0
dj-database-url==1.0.0
psycopg2-binary==2.9.2
gunicorn==20.1.0